requirements.txt
pyzmq==26.4.0
requests==2.32.3